+++ daemon/snap.go
@@ -28,6 +28,7 @@ import (
 	"strings"
 
 	"github.com/snapcore/snapd/client"
+	"github.com/snapcore/snapd/cmd"
 	"github.com/snapcore/snapd/logger"
 	"github.com/snapcore/snapd/overlord/assertstate"
 	"github.com/snapcore/snapd/overlord/snapstate"
@@ -286,7 +287,7 @@ func mapLocal(about aboutSnap) *client.Snap {
 	}
 	sort.Sort(bySnapApp(snapapps))
 
-	apps, err := client.AppInfosFromSnapAppInfos(snapapps)
+	apps, err := cmd.ClientAppInfosFromSnapAppInfos(snapapps)
 	if err != nil {
 		logger.Noticef("cannot get full app info: %v", err)
 	}
